+++ src/main/java/barch/mc_extended/ArFoMiToVi.java
@@ -4,6 +4,7 @@ import barch.mc_extended.Armor.ArmorMaterials;
 import barch.mc_extended.Entities.Entities;
 import barch.mc_extended.Foods.Foods;
 import barch.mc_extended.Minerals.Minerals;
+import barch.mc_extended.Misc.Misc;
 import barch.mc_extended.Tools.ToolMaterials;
 import barch.mc_extended.Villagers.Villagers;
 
@@ -13,15 +14,11 @@ public class ArFoMiToVi {
 
 
     public static final Minerals MINERALS = Minerals.INSTANCE;
-
     public static final Foods FOODS = Foods.INSTANCE;
-
+    public static final Misc MISC = Misc.INSTANCE;
     public static final ArmorMaterials ARMOR_MATERIALS = ArmorMaterials.INSTANCE;
-
     public static final ToolMaterials TOOL_MATERIALS = ToolMaterials.INSTANCE;
-
     public static final Villagers VILLAGERS = Villagers.INSTANCE;
-
     public static final Entities ENTITIES = Entities.INSTANCE;
 
 
@@ -29,6 +26,7 @@ public class ArFoMiToVi {
 
         MINERALS.RegisterAll();
         FOODS.RegisterAll();
+        MISC.RegisterAll();
 
         ARMOR_MATERIALS.RegisterAll();
         TOOL_MATERIALS.RegisterAll();