+++ src/main/resources/changelog.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.2] -2023-7-27
+## [0.2.3] - 2023-7-29
+### Added
+- trout fish
+- trout food item
+- cooked trout food item
+- trout bucket
+
+## [0.2.2] - 2023-7-27
 ### Added
 - lost entity
 
@@ -252,6 +259,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - deepslate ruby ore
 - ruby block
 
+[0.2.3]: https://github.com/BarchamMal/MC-Extended/commit/
+[0.2.2]: https://github.com/BarchamMal/MC-Extended/commit/b252f3e3333e4fd547f65bb483a0c393884aa9ca
 [0.2.1]: https://github.com/BarchamMal/MC-Extended/commit/c59a4616b5224cd8ddf78dc1fd22a8bbfe1cd345
 [0.2.0]: https://github.com/BarchamMal/MC-Extended/commit/b889162b3515908415f2afbe1a6deecee274651c
 [0.1.8]: https://github.com/BarchamMal/MC-Extended/commit/38d52ed944e0d9db02c2d9b154848b5325225452